Update by separating execute test
